enhance(ui):improve ui performance
