--- assets/css/blog-swiper.css
@@ -1,14 +1,0 @@
-/* .blog .swiper {
-  position: relative;
-  overflow: visible;
-}
-
-.blog .swiper-pagination-bullet {
-  background-color: green;
-  width: 15px;
-  height: 15px;
-}
-.blog .swiper-pagination {
-  bottom: -150px;
-  position: absolute !important;
-} */
