--- assets/css/package.css
@@ -1,18 +1,0 @@
-/* .swiper {
-  position: relative;
-  overflow: visible;
-}
-
-.swiper-pagination-bullet {
-  background-color: green;
-  width: 15px;
-  height: 15px;
-}
-.swiper--offset-small .swiper-pagination {
-  bottom: -50px !important;
-  position: absolute !important;
-}
-.swiper--offset-large .swiper-pagination {
-  bottom: -150px !important;
-  position: absolute !important;
-} */
