+++ components/VMG.vue
@@ -28,39 +28,5 @@
 </template>
 
 <script setup lang="ts">
-import aimBanner from "../assets/image/assests/OutGoal/aim-banner.png";
-import aimArrow from "../assets/image/assests/OutGoal/aim-arrow.svg";
-
-import visionBanner from "../assets/image/assests/OurVission/vision-banner.png";
-import visionIcon from "../assets/image/assests/OurVission/vison-icon.svg";
-
-import missionBanner from "../assets/image/assests/OurMission/mission-banner.png";
-import missionIcon from "../assets/image/assests/OurMission/mission-icon.png";
-
-const VGM = {
-  vision: {
-    topImg: visionBanner,
-    cardIcon: visionIcon,
-    carditle: "Our Vision",
-    cardDesc: ` The statement implies a willingness to help someone escape their daily routine by taking care of their travel plans. It suggests a travel agency or similar service. The statement implies a willingness to help someone escape their daily routine by taking care of their travel plans. It suggests a travel agency or similar service. The statement implies a willingness to help someone escape their daily routine by taking care of their travel plans. It suggests a travel agency or similar service. The statement implies a willingness to help someone escape their daily routine by taking care of their travel plans. It suggests a travel agency or similar service`,
-  },
-  mission: {
-    topImg: missionBanner,
-    cardIcon: missionIcon,
-    carditle: "Our Mission",
-    cardDesc: ` The statement implies a willingness to help someone escape their daily routine by taking care of their travel plans. It suggests a travel agency or similar service. The statement implies a willingness to help someone escape their daily routine by taking care of their travel plans. It suggests a travel agency or similar service. The statement implies a willingness to help someone escape their daily routine by taking care of their travel plans. It suggests a travel agency or similar service. The statement implies a willingness to help someone escape their daily routine by taking care of their travel plans. It suggests a travel agency or similar service`,
-  },
-  goal: {
-    topImg: aimBanner,
-    cardIcon: aimArrow,
-    carditle: "Our Goal",
-    cardDesc: ` The statement implies a willingness to help someone escape their daily routine by taking care of
-                        their travel plans. It suggests a travel agency or similar service. The statement implies a
-                        willingness to help someone escape their daily routine by taking care of their travel plans. It
-                        suggests a travel agency or similar service. The statement implies a willingness to help someone
-                        escape their daily routine by taking care of their travel plans. It suggests a travel agency or
-                        similar service. The statement implies a willingness to help someone escape their daily routine
-                        by taking care of their travel plans. It suggests a travel agency or similar service`,
-  },
-};
+import { VGM } from "../constant/index";
 </script>
